+++ siteapp/views.py
@@ -554,7 +554,7 @@ def start_app(appver, organization, user, folder, task, q, portfolio):
             user={"id": user.id, "username": user.username}
         )
         # Add deault deployments to system
-        deployment = Deployment(name="Design", description="Reference system archictecture design", system=system)
+        deployment = Deployment(name="Blueprint", description="Reference system archictecture design", system=system)
         deployment.save()
         deployment = Deployment(name="Dev", description="Development environment deployment", system=system)
         deployment.save()
@@ -566,16 +566,11 @@ def start_app(appver, organization, user, folder, task, q, portfolio):
         # Assign default control catalog and control profile
         # Use from App catalog settings
         try:
-            from pprint import pprint
-            pprint(project.root_task.module.app.catalog_metadata['params'])
             # Get default catalog key
-            params = project.root_task.module.app.catalog_metadata['params']
-            catalog_key = [p for p in params if p['id'] == 'default_catalog_key'][0]['value']
-            print("catalog_key", catalog_key)
+            parameters = project.root_task.module.app.catalog_metadata['parameters']
+            catalog_key = [p for p in parameters if p['id'] == 'catalog_key'][0]['value']
             # Get default profile/baseline
-            # params = project.root_task.module.app.catalog_metadata['params']
-            baseline_name = [p for p in params if p['id'] == 'default_baseline'][0]['value']
-            print("baseline_name", baseline_name)
+            baseline_name = [p for p in parameters if p['id'] == 'baseline'][0]['value']
             # Assign profile/baseline
             assign_results = system.root_element.assign_baseline_controls(user, catalog_key, baseline_name)
             # Log result if successful
@@ -605,7 +600,7 @@ def start_app(appver, organization, user, folder, task, q, portfolio):
                 user={"id": user.id, "username": user.username}
             )
 
-        # TODO: Assign default org params
+        # TODO: Assign default org parameters
 
         # Add user as the first admin.
         ProjectMembership.objects.create(
@@ -678,11 +673,6 @@ def project(request, project):
     #         if stage == project.lifecycle_stage[1]:
     #             break
 
-    # Get all of the discussions the user is participating in as a guest in this project.
-    # Meaning, I'm not a member, but I still need access to certain tasks and
-    # certain questions within those tasks.
-    discussions = list(project.get_discussions_in_project_as_guest(request.user))
-
     # Pre-load the answers to project root task questions and impute answers so
     # that we know which questions are suppressed by imputed values.
     root_task_answers = project.root_task.get_answers().with_extended_info()
@@ -738,7 +728,7 @@ def project(request, project):
                 "invitations": [], # filled in below
                 "task": module_answers.task,
                 "can_start_new_task": False,
-                "discussions": [d for d in discussions if d.attached_to.task == module_answers.task],
+                "discussions": [] # no longer tracking discussions per question,
             }
 
         # Create a "question" record for the question itself it is is unanswered or if
@@ -826,25 +816,10 @@ def project(request, project):
         if "id" in doc:
             has_outputs = True
 
-    # Find any open invitations and if they are for particular modules,
-    # display them with the module.
-    other_open_invitations = []
-    for inv in Invitation.objects.filter(from_user=request.user, from_project=project, accepted_at=None, revoked_at=None).order_by('-created'):
-        if inv.is_expired():
-            continue
-        if inv.target == project:
-            into_new_task_question_id = inv.target_info.get("into_new_task_question_id")
-            if into_new_task_question_id:
-                if into_new_task_question_id in questions: # should always be True
-                    questions[into_new_task_question_id]["invitations"].append(inv)
-                    continue
-
-        # If the invitation didn't get put elsewhere, display in the
-        # other list.
-        other_open_invitations.append(inv)
-
     # Calculate approximate compliance as degrees to display
-    percent_compliant = project.system.controls_status_count['Implemented'] / len(project.system.control_implementation_as_dict)
+    percent_compliant = 0
+    if len(project.system.control_implementation_as_dict) > 0:
+        percent_compliant = project.system.controls_status_count['Implemented'] / len(project.system.control_implementation_as_dict)
     # Need to reverse calculation for displaying as per styles in .piechart class
     approx_compliance_degrees = 365 - ( 365 * percent_compliant )
     if approx_compliance_degrees > 358:
@@ -867,7 +842,7 @@ def project(request, project):
         "can_start_any_apps": can_start_any_apps,
 
         "title": project.title,
-        "open_invitations": other_open_invitations,
+        # "open_invitations": other_open_invitations,
         "send_invitation": Invitation.form_context_dict(request.user, project, [request.user]),
         "has_outputs": has_outputs,